requirements.txt

numpy~=1.24.4
opencv-python~=4.10.0.84
scikit-image~=0.20.0
matplotlib~=3.7.2

torchvision~=0.18.1
pillow~=10.4.0
tqdm~=4.66.5
tensorflow-gpu~=2.10.0
keras~=2.10.0
scikit-learn~=1.3.2